+++ src/main/java/io/iridium/qolhunters/mixin/betterdescriptions/MixinAbilitiesDescriptionsConfig.java
@@ -1,4 +1,4 @@
-package io.iridium.qolhunters.mixin.configs;
+package io.iridium.qolhunters.mixin.betterdescriptions;
 
 import io.iridium.qolhunters.config.QOLHuntersClientConfigs;
 import org.spongepowered.asm.mixin.Mixin;
@@ -6,13 +6,13 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(iskallia.vault.config.BingoConfig.class)
-public class MixinBingoConfig {
+@Mixin(iskallia.vault.config.AbilitiesDescriptionsConfig.class)
+public class MixinAbilitiesDescriptionsConfig {
 
     @Inject(at = @At("HEAD"), method = "getName()Ljava/lang/String;", cancellable = true, remap = false)
     private void getName(CallbackInfoReturnable<String> cir) {
 
-        if(QOLHuntersClientConfigs.BETTER_BINGO_DESCRIPTIONS.get()) cir.setReturnValue("iridium/bingo");
+        if(QOLHuntersClientConfigs.BETTER_ABILITIES_DESCRIPTIONS.get()) cir.setReturnValue("iridium/abilities_descriptions");
     }
 
 }
