+++ src/main/java/io/iridium/qolhunters/mixin/betterdescriptions/MixinBingoConfig.java
@@ -1,19 +1,18 @@
-package io.iridium.qolhunters.mixin.configs;
+package io.iridium.qolhunters.mixin.betterdescriptions;
 
-import io.iridium.qolhunters.QOLHunters;
 import io.iridium.qolhunters.config.QOLHuntersClientConfigs;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(iskallia.vault.config.AbilitiesDescriptionsConfig.class)
-public class MixinAbilitiesDescriptionsConfig {
+@Mixin(iskallia.vault.config.BingoConfig.class)
+public class MixinBingoConfig {
 
     @Inject(at = @At("HEAD"), method = "getName()Ljava/lang/String;", cancellable = true, remap = false)
     private void getName(CallbackInfoReturnable<String> cir) {
 
-        if(QOLHuntersClientConfigs.BETTER_ABILITIES_DESCRIPTIONS.get()) cir.setReturnValue("iridium/abilities_descriptions");
+        if(QOLHuntersClientConfigs.BETTER_BINGO_DESCRIPTIONS.get()) cir.setReturnValue("iridium/bingo");
     }
 
 }
