+++ src/main/java/io/iridium/qolhunters/mixin/betterdescriptions/MixinSkillDescriptionsConfig.java
@@ -1,4 +1,4 @@
-package io.iridium.qolhunters.mixin.configs;
+package io.iridium.qolhunters.mixin.betterdescriptions;
 
 import io.iridium.qolhunters.config.QOLHuntersClientConfigs;
 import org.spongepowered.asm.mixin.Mixin;
@@ -6,12 +6,13 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(iskallia.vault.config.MenuPlayerStatDescriptionConfig.class)
-public class MixinMenuPlayerStatDescriptionConfig {
+@Mixin(iskallia.vault.config.SkillDescriptionsConfig.class)
+public class MixinSkillDescriptionsConfig {
 
     @Inject(at = @At("HEAD"), method = "getName()Ljava/lang/String;", cancellable = true, remap = false)
     private void getName(CallbackInfoReturnable<String> cir) {
-        if(QOLHuntersClientConfigs.BETTER_STATS_DESCRIPTIONS.get()) cir.setReturnValue("iridium/menu_player_stat_description");
+
+        if(QOLHuntersClientConfigs.BETTER_TALENTS_EXPERTISE_RESEARCH_DESCRIPTIONS.get()) cir.setReturnValue("iridium/skill_descriptions");
     }
 
 }
