+++ src/main/resources/qolhunters.mixins.json
@@ -20,11 +20,11 @@
     "boosterpack.MixinBoosterPackSelectionScreen",
     "brazier.MixinMonolithRenderer",
     "cake.MixinCakeObjective",
-    "configs.MixinAbilitiesDescriptionsConfig",
-    "configs.MixinBingoConfig",
-    "configs.MixinMenuPlayerStatDescriptionConfig",
+    "betterdescriptions.MixinAbilitiesDescriptionsConfig",
+    "betterdescriptions.MixinBingoConfig",
+    "betterdescriptions.MixinMenuPlayerStatDescriptionConfig",
     "configs.MixinModConfigs",
-    "configs.MixinSkillDescriptionsConfig",
+    "betterdescriptions.MixinSkillDescriptionsConfig",
     "godaltars.MixinGodAltarOverlay",
     "keybinds.MixinBountyScreen",
     "keybinds.MixinForgeRecipeContainerScreen",
